requirements.txt
python-telegram-bot>=20,<22
pillow>=10
python-dotenv>=1.0
